Modify shape_importer to parse shapefiles with polylines between stops instead of for entire trips
Review at http://codereview.appspot.com/8710

This change enables the shape importer to handle shapefiles
where there are separate polylines between every station on every line.

Some trips on some lines may skip stops if they are express trips.  So sometimes
we have to piece together several polylines to get from one station to the next
in a trip.  So to do all this, I implemented a version of the A* algorithm to
search for the shortest path between two stations, traveling over the graph of
all the polylines in the shape file.

Another wrinkle is that input shapefiles may not always be complete.  In order
to provide a way to manually patch things up, I added a flag to shape_importer
to allow passing in an extra shapes.txt file (in GTFS format) to augment the
polylines in the shapefile.

Along the way I added some debugging messages and a --verbosity flag to control
their output.

One tiny and only tangentially related change is to
gtfsscheduleviewer/files/index.html, adding lat/long to the pop-up bubble for
stations in the schedule viewer.  (Useful for debugging.)


git-svn-id: https://googletransitdatafeed.googlecode.com/svn/trunk@948 705c8f7f-ad29-0410-9695-93dd3b14b492