empty start_date/end_date warning
Fix for http://code.google.com/p/googletransitdatafeed/issues/detail?id=41

It is okay to have both start_date and end_date set or both None. This change
makes sure a problem is reported when either is set but not a valid date. It
drops the check for a directly created ServicePeriod with only start_date or
end_date set because it would duplicate the MissingColumn error.

Review at http://codereview.appspot.com/157070


git-svn-id: https://googletransitdatafeed.googlecode.com/svn/trunk@1171 705c8f7f-ad29-0410-9695-93dd3b14b492